analysis/requirements.txt
# Computational Model Dependencies
# For sequence clustering and behavioral analysis

# Core scientific computing
numpy>=1.24.0
scipy>=1.10.0
pandas>=2.0.0

# Time series clustering
tslearn>=0.6.2
dtaidistance>=2.3.10

# Machine learning
scikit-learn>=1.3.0

# Statistical analysis
statsmodels>=0.14.0

# Database connectivity
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0  # PostgreSQL adapter

# Embeddings (optional, if not using external API)
sentence-transformers>=2.2.0

# Utilities
python-dotenv>=1.0.0
tqdm>=4.65.0  # Progress bars

